+++ src/Foundation/Shipping/PaymentDependentShippingFee.php
@@ -25,6 +25,8 @@ use Vanilo\Adjustments\Support\IsNotIncluded;
 
 class PaymentDependentShippingFee implements Adjuster
 {
+    public const ALIAS = 'payment_dependent_shipping_fee';
+
     use HasWriteableTitleAndDescription;
     use IsLockable;
     use IsNotIncluded;