+++ Changelog.md
@@ -47,6 +47,8 @@
 - Added the `Merchant` interface
 - Added the `DefaultTaxCalculator` class - calculates simply by rate
 - Added the `CalculateTaxes` listener to cart update and shipping address change events
+- Added the `AdjusterAliases` class that for decoupling FQCNs from the database
+- Added automatic mapping of adjuster FQCN <-> aliases when saving an adjustment into the DB and when calling the `getAdjuster()` method
 - BC: Added the `?CheckoutSubject` return type to the `getCart()` method of the `Checkout` interface
 - BC: Changed `Checkout::getShippingAddress()` return type to be nullable
 - BC: Added the void return type to `Checkout::setShippingAddress()`